@JsonIgnore getDecoratedClass() in projection
Oliver Gierke resolved:
Looks like this is caused by proxies created through Spring's ProxyFactory now also implementing a new DecoratingProxy, which exposes getDecoratedClass() and our TargetAware interface that's in place to mask the artificial proxy properties doesn't mask that newly introduced attribute. I can adapt our TargetAware accordingly.
A temporary workaround would be to redeclare the getDecoratedClass() method on your projection interface and annotate it with @JsonIgnore.

Dmitry Stolbov added a comment:
This decide with @JsonIgnore getDecoratedClass() is worked.
But it don't work for query methods of Repository, which return Projection:
Unable to locate Attribute with the the given name decoratedClass on this ManagedType ru.inkontext.domain.Person
I do not use this method in my project and @Ignore tests in example tests.
Thank you.